backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# NLP Libraries
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.2.2
textstat==0.7.3
language-tool-python==2.7.1

# ML/Data
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
torch>=2.0.0

# API Integration
google-generativeai==0.3.0
httpx==0.25.2
python-dotenv==1.0.0
aiohttp==3.9.1

# Caching
diskcache==5.6.3

# Export/Visualization
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.7
plotly==5.18.0
python-docx==1.1.0
PyPDF2==3.0.1

# Utilities
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# WebSocket support
websockets==12.0
python-socketio==5.10.0